Renames `platform-support` to `polyfill-support`
