+++ packages/lit-element/CHANGELOG.md
@@ -55,7 +55,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [Breaking] Removed build support for TypeScript 3.4.
 - [Breaking] Decorators are no longer exported from the `lit-element` module. Instead, import any decorators you use from `lit-element/decorators/*`.
 - [Breaking] `lit-html` has been updated to 2.x.
-- [Breaking] Support for running in older browsers has been removed from the default configuration. Import the `platform-support` module to support Shady DOM. Note also that Lit parts inside `<style>` elements are no longer supported.
+- [Breaking] Support for running in older browsers has been removed from the default configuration. Import the `polyfill-support` module to support Shady DOM. Note also that Lit parts inside `<style>` elements are no longer supported.
 - [Breaking] For simplicity, `requestUpdate` no longer returns a Promise. Instead await the `updateComplete` Promise.
 - [Breaking] The type of the `css` function has been changed to `CSSResultGroup` and is now the same as `LitElement.styles`. This avoids the need to cast the `styles` property to `any` when a subclass sets `styles` to an Array and its super class set a single value (or visa versa).
 - For efficiency, the `css` function now maintains a cache and will use a cached value if available when the same style text is requested.
