+++ packages/lit-element/src/test/polyfill-support/lit-element-bundle_test.html
@@ -4,7 +4,7 @@
     <!-- Loads polyfills but does not force settings... -->
     <meta name="manual-polyfills" />
     <script src="../../../node_modules/@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-    <script src="../../../platform-support.js"></script>
+    <script src="../../../polyfill-support.js"></script>
     <script src="../../../lit.min.js"></script>
   </head>
