+++ packages/lit-element/src/test/polyfill-support/lit-element_html-test.ts
@@ -11,7 +11,7 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-import '../../platform-support.js';
+import '../../polyfill-support.js';
 import '../lit-element_styling_test.js';
 import '../lit-element_styling_apply_test.js';
 import '../lit-element_test.js';
