+++ packages/lit-html/rollup.config.js
@@ -37,11 +37,11 @@ export default litProdConfig({
     'static',
     'hydrate',
     'private-ssr-support',
-    'platform-support',
+    'polyfill-support',
   ],
   bundled: [
     {
-      file: 'platform-support',
+      file: 'polyfill-support',
     },
   ],
 });
