+++ packages/lit-html/src/polyfill-support.ts
