+++ packages/lit-html/src/test/polyfill-support/lit-html-no-wc_html-test.ts
@@ -11,7 +11,7 @@
  * subject to an additional IP rights grant found at
  * http://polymer.github.io/PATENTS.txt
  */
-import '../../platform-support.js';
+import '../../polyfill-support.js';
 import {html, render} from '../../lit-html.js';
 import {assert} from '@esm-bundle/chai';
 import {ShadyRenderOptions} from '../test-utils/shadow-root.js';
