+++ packages/reactive-element/.gitignore
@@ -3,6 +3,6 @@
 /test/
 /css-tag.*
 /decorators.*
-/platform-support.*
+/polyfill-support.*
 /reactive-element.*
 /reactive-controller.*
